Change to typescript
