+++ babel.config.json
@@ -1,0 +1,8 @@
+{
+  "presets": [
+    ["@babel/preset-env", {
+      "targets": "> 0.25%, not dead"
+    }],
+    "@babel/preset-typescript"
+  ]
+}
