+++ package.json
@@ -3,8 +3,9 @@
   "version": "1.0.2",
   "description": "Responsive map according to image size",
   "main": "dist/index.js",
+  "types": "dist/index.d.ts",
   "scripts": {
-    "build": "webpack",
+    "build": "webpack --config webpack.config.js",
     "prepublishOnly": "npm run build",
     "test": "jest --verbose"
   },
@@ -16,6 +17,7 @@
     "responsive",
     "image",
     "map",
+    "typescript",
     "javascript"
   ],
   "author": "geongu.park",
@@ -28,9 +30,14 @@
     "@babel/cli": "^7.24.8",
     "@babel/core": "^7.25.2",
     "@babel/preset-env": "^7.25.3",
+    "@babel/preset-typescript": "^7.24.7",
+    "@types/jest": "^29.5.12",
     "babel-loader": "^9.1.3",
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
+    "ts-jest": "^29.2.4",
+    "ts-loader": "^9.5.1",
+    "typescript": "^5.5.4",
     "webpack": "^5.93.0",
     "webpack-cli": "^5.1.4"
   }
