+++ webpack.config.js
@@ -1,24 +1,29 @@
 const path = require('path');
 
 module.exports = {
-  entry: './src/index.js',
+  mode: 'production',
+  entry: './src/index.ts',
   output: {
     filename: 'index.js',
     path: path.resolve(__dirname, 'dist'),
     library: 'ResponsiveImageMap',
     libraryTarget: 'umd',
     globalObject: 'this',
+    umdNamedDefine: true
   },
   module: {
     rules: [
       {
-        test: /\.js$/,
+        test: /\.ts$/,
+        use: [
+          'babel-loader',
+          'ts-loader',
+        ],
         exclude: /node_modules/,
-        use: {
-          loader: 'babel-loader',
-        },
       },
     ],
   },
-  mode: 'production',
+  resolve: {
+    extensions: ['.ts', '.js']
+  }
 };
